Binary tree using Linked Nodes and Arrays.
An interface BinaryTree that includes two methods to perform insertion and level order traversal (ie., breadth-first traversal). A class ArrayBinaryTree that implements the interface so as to perform the binary tree operations in an array. Another class LinkedBinaryTree that implements the above interface so as to perform the binary tree operations with dynamically created nodes. Invoke these methods from the main class to create ‘n’ nodes. Root node stores “0”. The left child of any node appends “0” to the content of the parent node and the right child of any node should append “1” to the content of the parent node.