Treemaps and other collection class's objects.
A text file contains several lines of sentences as follows. Each line contains a sequence of words and their corresponding parts-of-speech within a pair of parentheses. Here, AUX, N,V,ADJ,DET refer to auxiliary verb, noun, verb, adjective and determiner respectively. Read them and create a TreeMap to store each parts-of-speech as the key and the list of words which have been tagged with that parts-of-speech as the value. E.g., N={Brown, coffee, Swift, car, drive}	
Did Brown drink coffee (AUX N V N)
Coffee is brown (N AUX ADJ)
Swift is a car (N AUX DET N)
Car is brown (N AUX ADJ)
Brown had a swift drive (N AUX DET ADJ N)
Did Brown drive (AUX N V)
Read an input phrase as the one given below and store all possible parts-of-speech for each word in a 2-D ragged array. Read the user-given tag sequence for the phrase and check if it is an acceptable tagging or not, based on the 2-D ragged array. Print the map, array and the acceptance status.