Command Line Arguments. (Using Bank example.)
The details of a list of transactions carried out on a single day for a particular bank are given as command line inputs in the order of customer name, total amount, transacted amount. Read these inputs. If the transacted amount is negative, it can be considered as withdrawal. If it is positive, it can be considered as deposit. For withdrawal, either if the total amount is less than the |transacted amount| or if the withdrawal amount exceeds the maximum limit of 25000, display the message, “Failed Transaction”. Withdrawal charges are as follows. 
If the withdrawal amount <=500, charge=5
If the withdrawal amount <=1000, charge=8
If the withdrawal amount >100 and <=5000, charge=10
If the withdrawal amount >5000 and <=15000, charge =12
If the withdrawal amount >15000 and <=25000, charge=15
For every customer, print the name and his balance amount after the transaction.