Encoding by replacing words w/ numbers.
Define a class ‘Encoding’ with the instance variables – inputtext (a string), SA (a String array to store the individual words of the inputtext) and IA (an integer array to store the encoded values). [Note: Each object of this class will have a copy of inputtext, SA and IA]. Define a constructor, to split the inputtext into words and store them in the String array, SA. In the main class, read ‘n’ such input texts and store their details in an array of ‘n’ objects. Define a method, sort() in the ‘Encoding’ class to sort the individual words of all the ‘n’ input texts, in alphabetical order. Define another method, encodeText() to encode the individual words of each text based on the index of the word in the sorted list and store them in the integer array. Overload this method to make the integer array of each inputtext to have equal number of elements by padding with the value, -1. 
Invoke all these methods from the main class.